convertdocx/requirements.txt
beautifulsoup4==4.13.4
lxml==5.4.0
markdown2==2.5.3
python-docx==1.2.0
PyYAML==6.0.2
soupsieve==2.7
typing_extensions==4.14.0
